fix type

+++ sstd/src/file/file_operations.hpp
@@ -1,7 +1,8 @@
 #pragma once
 
+#include <stddef.h>
 #include "../typeDef.h"
 
 namespace sstd{
-    bool file_size(uint64& ret_size, const char* path);
+    bool file_size(size_t& ret_size, const char* path);
 }
